refactor(types): switch
